fixed FIFA program and #6 on the midterm. Had floats for numbers that should have been int because I couldn't figure out how to divide integers to get a float in return.

+++ Week6/FIFAProgram.c
@@ -11,14 +11,14 @@
 // initialize structure for all team statistics
 struct teamStats
 {
-    float shotsOnGoalie;
-    float goalsScoredAgainst;
-    float goalieGamesWon;
-    float goalieGamesLost;
-    float scoringAttempts;
-    float goalsScoredFor;
-    float teamGamesWon;
-    float teamGamesLost;
+    int shotsOnGoalie;
+    int goalsScoredAgainst;
+    int goalieGamesWon;
+    int goalieGamesLost;
+    int scoringAttempts;
+    int goalsScoredFor;
+    int teamGamesWon;
+    int teamGamesLost;
 };
 
 //main function begin
@@ -61,20 +61,20 @@ int main()
 void defenseStats(struct teamStats blueTeam)
 {
 
-    printf("Team Defense\n__________\n\nShots on Goalie: %f\nGoals Scored Against: %f\nGoalie Wins: %f\nGoalie Losses: %f\n\n", 
+    printf("Team Defense\n__________\n\nShots on Goalie: %i\nGoals Scored Against: %i\nGoalie Wins: %i\nGoalie Losses: %i\n\n", 
         blueTeam.shotsOnGoalie, blueTeam.goalsScoredAgainst, blueTeam.goalieGamesWon, blueTeam.goalieGamesLost);
 
     float goalieWinPercentage = 0.0;
     float goalieSavePercentage = 0.0;
     float goalieSaveRatio = 0.0;
-    float goalieSaves = 0;
+    int goalieSaves = 0;
 
-    goalieWinPercentage = blueTeam.goalieGamesWon / (blueTeam.goalieGamesWon + blueTeam.goalieGamesLost);
+    goalieWinPercentage = (float)blueTeam.goalieGamesWon / ((float)blueTeam.goalieGamesWon + (float)blueTeam.goalieGamesLost);
     goalieSaves = blueTeam.shotsOnGoalie - blueTeam.goalsScoredAgainst;
-    goalieSavePercentage = goalieSaves / blueTeam.shotsOnGoalie;
-    goalieSaveRatio = blueTeam.shotsOnGoalie / goalieSaves;
+    goalieSavePercentage = (float)goalieSaves / (float)blueTeam.shotsOnGoalie;
+    goalieSaveRatio = (float)blueTeam.shotsOnGoalie / (float)goalieSaves;
 
-    printf("Win Percentage: %f\nSaves: %f\nSave Percentage: %f\nSave Ratio: %f\n\n\n", 
+    printf("Win Percentage: %.3f\nSaves: %i\nSave Percentage: %.3f\nSave Ratio: %.3f\n\n\n", 
         goalieWinPercentage, goalieSaves, goalieSavePercentage, goalieSaveRatio);
 
 }
@@ -98,17 +98,17 @@ void defenseStats(struct teamStats blueTeam)
 //******************************************************************************************
 void offenseStats(struct teamStats blueTeam)
 {
-    printf("Team Offense\n____________\n\nScoring Attempts: %f\nGoals Scored For: %f\nTeam Games Won: %f\nTeam Games Lost: %f\n\n", 
+    printf("Team Offense\n____________\n\nScoring Attempts: %i\nGoals Scored For: %i\nTeam Games Won: %i\nTeam Games Lost: %i\n\n", 
         blueTeam.scoringAttempts, blueTeam.goalsScoredFor, blueTeam.teamGamesWon, blueTeam.teamGamesLost);
 
     float scoringPercentage = 0.0;
     float scoringRatio = 0.0;
     float gamesWonPercentage = 0.0;
 
-    scoringPercentage = (blueTeam.scoringAttempts - blueTeam.goalsScoredFor) / blueTeam.scoringAttempts;
-    scoringRatio = blueTeam.scoringAttempts/ blueTeam.goalsScoredFor;
-    gamesWonPercentage = blueTeam.teamGamesWon / (blueTeam.teamGamesWon + blueTeam.teamGamesLost);
+    scoringPercentage = ((float)(blueTeam.scoringAttempts - blueTeam.goalsScoredFor)) / (float)blueTeam.scoringAttempts;
+    scoringRatio = (float)blueTeam.scoringAttempts/ (float)blueTeam.goalsScoredFor;
+    gamesWonPercentage = (float)blueTeam.teamGamesWon / ((float)(blueTeam.teamGamesWon + blueTeam.teamGamesLost));
 
-    printf("Scoring Percentage: %f\nScoring Ratio: %f\nGames Won Percentage: %f\n\n", 
+    printf("Scoring Percentage: %.3f\nScoring Ratio: %.3f\nGames Won Percentage: %.3f\n\n", 
         scoringPercentage, scoringRatio, gamesWonPercentage);
 }
